% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{add_logo}
\alias{add_logo}
\title{Logo}
\usage{
add_logo(hc, path, top = 80L, left = 10L, width = 100L, height = 100L)
}
\arguments{
\item{hc}{a highcharter chart}

\item{path}{path to logo}

\item{top}{top margin of logo}

\item{left}{left margin of logo}

\item{width}{width of logo}

\item{height}{height of logo}
}
\description{
Add logo to exported charts.
}
\examples{

library(naccho.viz)
library(highcharter)

    highchart() |>
      hc_xAxis(categories = month.abb) |>
      hc_add_series(name = "Tokyo", data = sample(1:12)) |>
      hc_exporting(
        enabled = TRUE, # always enabled
        filename = "custom-file-name"
        ) |>
      add_logo(
        "https://www.naccho.org/images/naccho-logo-banner.svg",
        width = 10L,
        height = 10L
      )

}
