% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{bar_pos_neg_theme}
\alias{bar_pos_neg_theme}
\title{NACCHO Bar Graph with Positive and Negative Values Theme}
\format{
\subsection{\code{bar_pos_neg_theme}}{

A list with a length of 12
}
}
\source{
adapted from the NACCHO basic theme
}
\usage{
bar_pos_neg_theme
}
\description{
A list containing NACCHO's bar graph theme with a plotline at "0"
}
\keyword{datasets}
