% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{bare}
\alias{bare}
\title{NACCHO Bare  Theme}
\format{
\subsection{\code{bare}}{

A list with a length of 12
}
}
\source{
adapted from the NACCHO style guide for bare graphs
}
\usage{
bare
}
\description{
A list containing NACCHO's bare highcharter theme for bare graphs
}
\keyword{datasets}
