% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{basic}
\alias{basic}
\title{NACCHO Basic Theme}
\format{
\subsection{\code{basic}}{

A list with a length of 12
}
}
\source{
adapted from the NACCHO minimal theme, add axis lines
}
\usage{
basic
}
\description{
A list containing NACCHO's basic highcharter theme
}
\keyword{datasets}
