% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{butterfly}
\alias{butterfly}
\title{NACCHO Bar Graph with Positive and Negative Values Theme}
\format{
\subsection{\code{butterfly}}{

A list with a length of 12
}
}
\source{
adapted from the NACCHO basic theme
}
\usage{
butterfly
}
\description{
A list containing NACCHO's bar graph theme with a plotline at "0"
}
\keyword{datasets}
