% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{create_cat_map}
\alias{create_cat_map}
\title{Create an accessible categorical map using the NACCHO theme in {highcharter}}
\usage{
create_cat_map(
  data,
  fips_column,
  selected_value,
  include_islands = FALSE,
  my_colors = naccho.viz::colors,
  my_click = NA,
  title_text,
  subtitle_text,
  caption_text,
  legend_text = NA,
  accessible_desc,
  tooltip = "{point.value}",
  legend_reverse = FALSE,
  enable_legend = TRUE
)
}
\arguments{
\item{data}{the data frame to create the map.}

\item{fips_column}{column of a dataframe with the Federal Information Processing System (FIPS) Codes for States and Counties, in quotes.}

\item{selected_value}{column of a dataframe with the values to visualize in the map, in quotes.}

\item{include_islands}{boolean, TRUE or FALSE, where FALSE removes island areas. Defaults to TRUE.}

\item{my_colors}{vector; colors to supply to the map, as hex values.}

\item{my_click}{JS function; use to add a click event, requires a JS function, JS(), Defaults to NA.}

\item{title_text}{character string; title of the map.}

\item{subtitle_text}{character string; subtitle to appear underneath the title, typically a date or year.}

\item{caption_text}{character string; source line and data notes to appear under the map.}

\item{legend_text}{character string; title of the legend. Defaults to NA.}

\item{accessible_desc}{character string; alternative text description of the figure for screen readers.}

\item{tooltip}{character string; text and values to appear in the tooltip, use {point.xx} to call a value from the data frame, where xx is a specific column.}

\item{legend_reverse}{boolean, TRUE or FALSE, where TRUE reverses the legend. Defaults to FALSE.}

\item{enable_legend}{boolean, TRUE or FALSE, where TRUE enables the legend. Defaults to TRUE.}
}
\value{
A categorical map
}
\description{
Create an accessible categorical map using the NACCHO theme in {highcharter}
}
