% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{create_grouped_line}
\alias{create_grouped_line}
\title{NACCHO Grouped Line Graph}
\usage{
create_grouped_line(
  data,
  x_variable,
  y_variable,
  x_group,
  my_colors = naccho.viz::naccho_colors,
  x_label,
  y_label,
  y_max = NULL,
  y_format = "{value:,.0f}",
  title_text = "",
  subtitle_text = "",
  caption_text = "",
  tooltip,
  legend_enable = TRUE,
  legend_title = "Category",
  select_theme = naccho.viz::naccho_minimal_theme,
  allow_export = TRUE,
  accessible_desc = ""
)
}
\arguments{
\item{data}{the data frame to create the map.}

\item{x_variable}{object; from the data frame, ideally a character.}

\item{y_variable}{numeric; from the data frame.}

\item{x_group}{character object; from the data frame.}

\item{my_colors}{vector or character string; one character or a vector of characters, which must be defined hex values. Defaults to naccho.viz::naccho_colors.}

\item{x_label}{character string; label for the x-axis.}

\item{y_label}{character string; label for the y-axis.}

\item{y_max}{numeric; the maximun value of the y-axis, defaults to 100.}

\item{y_format}{character string; format for the y-axis numbers, defaults to "{value:,.0f}". This is used to adjust decimals or add characters after the number. Examples follow. "{value}\%" add a \% sign and does not specify decimals. "{value:.1f}" specifies one decimal point. "{value:,.0f}" specifies one decimal and adds a comma every three values. "{value}\%" adds a percent sign. You can also add words to the value by doing the following: "{value} words here".}

\item{title_text}{character string; title of the visualization. Defaults to nothing.}

\item{subtitle_text}{character string; subtitle to appear underneath the title. Defaults to nothing.}

\item{caption_text}{character string; source and data notes to appear underneath the figure. Defaults to nothing.}

\item{tooltip}{character string; text and values to appear in the tooltip, use {point.xx} to call a value from the data frame, where xx is a specific column.}

\item{legend_enable}{boolean, TRUE or FALSE, where TRUE enables the legend. Defaults to FALSE.}

\item{legend_title}{character string; title of the legend, defaults to nothing unless defined.}

\item{select_theme}{hc_theme, must be a defined hc_theme. Options from within this package include naccho.viz::naccho_minimal_theme and naccho_normal.}

\item{allow_export}{boolean, TRUE or FALSE, where true allows the visualization to be exported.}

\item{accessible_desc}{character string; alternative text description of the figure for screen readers.}
}
\value{
A line graph from highcharter.
}
\description{
NACCHO Grouped Line Graph
}
\examples{
test_data <- base::data.frame(
  year = c("2019", "2022", "2019", "2022"),
  value = c(150, 200, 20, 30),
  bender = c("Fire Bender", "Fire Bender", "Water Bender", "Water Bender")
)
test_data <-
within(test_data, bender <- factor(bender, levels=c("Water Bender", "Fire Bender")))
test_data <- test_data \%>\%
  dplyr::arrange(bender)
create_grouped_line(
  data = test_data,
  x_variable = "year",
  y_variable = "value",
  x_group = "bender",
  y_label = "Number",
  x_label = "Year",
  y_max = 250,
  legend_enable = TRUE,
  legend_title = "Bender Type",
  tooltip =
    "Bender: {point.bender} <br>
 Value: {point.value}"
)

}
