% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.R
\name{create_icon_item}
\alias{create_icon_item}
\title{NACCHO Icon Item Chart}
\usage{
create_icon_item(
  data,
  y_variable,
  item_group,
  marker_icon = "person",
  title_text = "",
  subtitle_text = "",
  caption_text = "",
  caption_align = "center",
  legend_enable = FALSE,
  legend_title = "",
  tooltip,
  select_theme = naccho.viz::minimal,
  allow_export = TRUE,
  accessible_desc = "",
  color_one = naccho.viz::teal,
  color_two = naccho.viz::green,
  color_thee = naccho.viz::navy,
  color_four = naccho.viz::pink,
  color_five = naccho.viz::terracotta
)
}
\arguments{
\item{data}{the data frame to create the map.}

\item{y_variable}{numeric; from the data frame.}

\item{item_group}{character object; from the data frame.}

\item{marker_icon}{character object that corresponds to fontawesome icons; defaults to person.}

\item{title_text}{character string; title of the visualization. Defaults to nothing.}

\item{subtitle_text}{character string; subtitle to appear underneath the title. Defaults to nothing.}

\item{caption_text}{character string; source and data notes to appear underneath the figure. Defaults to nothing.}

\item{caption_align}{character string; options include "left", "center", or "right" where the default is "center".}

\item{legend_enable}{boolean, TRUE or FALSE, where TRUE enables the legend. Defaults to FALSE.}

\item{legend_title}{character string; title of the legend, defaults to nothing unless defined.}

\item{tooltip}{character string; text and values to appear in the tooltip, use {point.xx} to call a value from the data frame, where xx is a specific column.}

\item{select_theme}{hc_theme, must be a defined hc_theme. Options from within this package include naccho.viz::minimal and naccho_normal.}

\item{allow_export}{boolean, TRUE or FALSE, where true allows the visualization to be exported.}

\item{accessible_desc}{character string; alternative text description of the figure for screen readers.}

\item{color_one}{hex value or named color, first color.}

\item{color_two}{hex value or named color, second color.}

\item{color_thee}{hex value or named color, third color.}

\item{color_four}{hex value or named color, fourth color.}

\item{color_five}{hex value or named color, fifth color.}
}
\value{
an item chart with a font awesome icon.
}
\description{
NACCHO Icon Item Chart
}
\examples{
test_data <- base::data.frame(
  value = c(6, 4),
  bender = c("Fire Bender", "Water Bender")
)
create_icon_item(
  data = test_data,
  y_variable = "value",
  item_group = "bender",
  legend_enable = TRUE,
  legend_title = "Bender Type",
  tooltip =
    "{point.value}\%"
)
}
