% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{create_single_bar}
\alias{create_single_bar}
\title{NACCHO Single Bar Chart}
\usage{
create_single_bar(
  data,
  type = "column",
  x_variable,
  y_variable,
  tick_amount = 6,
  color_bars = FALSE,
  y_label = "",
  y_max = 100,
  y_format = "{value:,.0f}",
  my_colors = naccho.viz::colors,
  title_text = "",
  subtitle_text = "",
  caption_text = "",
  caption_align = "center",
  legend_enable = FALSE,
  legend_title = "",
  tooltip = NULL,
  select_theme = naccho.viz::minimal,
  allow_export = TRUE,
  accessible_desc = ""
)
}
\arguments{
\item{data}{the data frame to create the map.}

\item{type}{character string; must be "column" or "bar", where "column" displays vertically and "bar" displays horizontally.}

\item{x_variable}{object; from the data frame, ideally a character.}

\item{y_variable}{numeric; from the data frame.}

\item{tick_amount}{numeric; sets the number of ticks, defaults to 6.}

\item{color_bars}{boolean, TRUE or FALSE, where TRUE results in each bar being a different color.}

\item{y_label}{character string; label for the y-axis.}

\item{y_max}{numeric; the maximun value of the y-axis, defaults to 100.}

\item{y_format}{character string; format for the y-axis numbers, defaults to "{value:,.0f}". This is used to adjust decimals or add characters after the number. Examples follow. "{value}\%" add a \% sign and does not specify decimals. "{value:.1f}" specifies one decimal point. "{value:,.0f}" specifies one decimal and adds a comma every three values. "{value}\%" adds a percent sign. You can also add words to the value by doing the following: "{value} words here".}

\item{my_colors}{vector or character string; one character or a vector of characters, which must be defined hex values. Defaults to naccho.viz::colors.}

\item{title_text}{character string; title of the visualization. Defaults to nothing.}

\item{subtitle_text}{character string; subtitle to appear underneath the title. Defaults to nothing.}

\item{caption_text}{character string; source and data notes to appear underneath the figure. Defaults to nothing.}

\item{caption_align}{character string; options include "left", "center", or "right" where the default is "center".}

\item{legend_enable}{boolean, TRUE or FALSE, where TRUE enables the legend. Defaults to FALSE.}

\item{legend_title}{character string; title of the legend, defaults to nothing unless defined.}

\item{tooltip}{character string; text and values to appear in the tooltip, use {point.xx} to call a value from the data frame, where xx is a specific column.}

\item{select_theme}{hc_theme, must be a defined hc_theme.}

\item{allow_export}{boolean, TRUE or FALSE, where true allows the visualization to be exported.}

\item{accessible_desc}{character string; alternative text description of the figure for screen readers.}
}
\value{
A column or bar chart from highcharter.
}
\description{
Create an accessible single bar chart (vertical or horizontal) using a NACCHO theme and the \{highcharter\} package.
}
\examples{
test_data <- data.frame(
  year = c("2019", "2022"),
  value = c(150, 200)
)

create_single_bar(
  data = test_data,
  x_variable = "year",
  y_variable = "value",
  y_label = "Number",
  y_max = 250,
  tooltip = "{point.value}"
)

}
