% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.R
\name{create_single_item}
\alias{create_single_item}
\title{NACCHO Single Item Chart}
\usage{
create_single_item(
  data,
  y_variable,
  item_group,
  my_colors = naccho.viz::naccho_colors,
  title_text = "",
  subtitle_text = "",
  caption_text = "",
  caption_align = "center",
  legend_enable = FALSE,
  legend_title = "",
  tooltip,
  select_theme = naccho.viz::minimal,
  allow_export = TRUE,
  accessible_desc = ""
)
}
\arguments{
\item{data}{the data frame to create the map.}

\item{y_variable}{numeric; from the data frame.}

\item{item_group}{character object; from the data frame.}

\item{my_colors}{vector or character string; one character or a vector of characters, which must be defined hex values. Defaults to naccho.viz::naccho_colors.}

\item{title_text}{character string; title of the visualization. Defaults to nothing.}

\item{subtitle_text}{character string; subtitle to appear underneath the title. Defaults to nothing.}

\item{caption_text}{character string; source and data notes to appear underneath the figure. Defaults to nothing.}

\item{caption_align}{character string; options include "left", "center", or "right" where the default is "center".}

\item{legend_enable}{boolean, TRUE or FALSE, where TRUE enables the legend. Defaults to FALSE.}

\item{legend_title}{character string; title of the legend, defaults to nothing unless defined.}

\item{tooltip}{character string; text and values to appear in the tooltip, use {point.xx} to call a value from the data frame, where xx is a specific column.}

\item{select_theme}{hc_theme, must be a defined hc_theme. Options from within this package include naccho.viz::minimal and naccho_normal.}

\item{allow_export}{boolean, TRUE or FALSE, where true allows the visualization to be exported.}

\item{accessible_desc}{character string; alternative text description of the figure for screen readers.}
}
\value{
A column or bar chart from highcharter.
}
\description{
Create an accessible stacked bar chart (vertical or horizontal) using a NACCHO theme and the {highcharter} package.
}
\examples{
test_data <- base::data.frame(
  value = c(65, 11, 24),
  bender = c("Fire Bender", "Water Bender", "Air Bender")
)
test_data <-
 within(test_data, bender <- factor(bender, levels=c("Air Bender", "Water Bender", "Fire Bender")))
test_data <- test_data \%>\%
  dplyr::arrange(bender)
create_single_item(
  data = test_data,
  y_variable = "value",
  item_group = "bender",
  legend_enable = TRUE,
  legend_title = "Bender Type",
  tooltip =
    "{point.value}\%"
)
}
