% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{create_tick_bar}
\alias{create_tick_bar}
\title{NACCHO Tick Bar Chart}
\usage{
create_tick_bar(
  data,
  type = "bar",
  x_variable,
  y_variable,
  x_group,
  size = 400,
  y_format = "{value:,.0f}",
  my_colors = c(naccho.viz::naccho_offwhite, naccho.viz::naccho_teal),
  title_text = "",
  subtitle_text = "",
  caption_text = "",
  caption_align = "center",
  enable_labels = TRUE,
  select_theme = naccho.viz::naccho_minimalTick_theme,
  reverse = FALSE,
  tooltip,
  allow_export = TRUE,
  accessible_desc = ""
)
}
\arguments{
\item{data}{the data frame to create the map.}

\item{type}{character string; must be "column" or "bar", where "column" displays vertically and "bar" displays horizontally.}

\item{x_variable}{object; from the data frame, ideally a character.}

\item{y_variable}{numeric; from the data frame.}

\item{x_group}{character object; from the data frame.}

\item{size}{numeric; controls width of figure, defaults to 400.}

\item{y_format}{character string; format for the y-axis numbers, defaults to "{value:,.0f}". This is used to adjust decimals or add characters after the number. Examples follow. "{value}\%" add a \% sign and does not specify decimals. "{value:.1f}" specifies one decimal point. "{value:,.0f}" specifies one decimal and adds a comma every three values. "{value}\%" adds a percent sign. You can also add words to the value by doing the following: "{value} words here".}

\item{my_colors}{vector or character string; one character or a vector of characters, which must be defined hex values. Defaults to naccho.viz::naccho_colors.}

\item{title_text}{character string; title of the visualization. Defaults to nothing.}

\item{subtitle_text}{character string; subtitle to appear underneath the title. Defaults to nothing.}

\item{caption_text}{character string; source and data notes to appear underneath the figure. Defaults to nothing.}

\item{caption_align}{character string; options include "left", "center", or "right" where the default is "center".}

\item{enable_labels}{boolean, TRUE or FALSE, where FALSE removes labels.}

\item{select_theme}{select a theme, defaults to naccho.viz::naccho_minimalTick_theme.}

\item{reverse}{boolean, TRUE or FALSE, where TRUE reverses the figure.}

\item{tooltip}{character string; text and values to appear in the tooltip, use {point.xx} to call a value from the data frame, where xx is a specific column.}

\item{allow_export}{boolean, TRUE or FALSE, where true allows the visualization to be exported.}

\item{accessible_desc}{character string; alternative text description of the figure for screen readers.}
}
\value{
A ticked chart from highcharter.
}
\description{
Create an accessible ticked bar chart (horizontal) using a NACCHO theme and the {highcharter} package.
}
\examples{
test_data <- base::data.frame(
  action = c("Expanded", "Reduced", "No Change", "Expanded", "Reduced", "No Change"),
  value = c(6.5, 1.1, 2.4, 3.5, 8.9, 7.6),
  display = c(1, 1, 1, 0, 0, 0)
)

test_data <-
within(test_data, action <- factor(action, levels=c("Expanded", "Reduced", "No Change")))

test_data <- test_data \%>\%
 dplyr::arrange(action)
create_tick_bar(
  data = test_data,
 x_variable = "action",
  y_variable = "value",
 x_group = "display",
 tooltip = "Value: {point.value} <br> "
)
}
