% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{green}
\alias{green}
\title{NACCHO Green}
\format{
\subsection{\code{green}}{

A single vector as a hex code
}
}
\source{
adapted from the NACCHO style guide for greater accessibility
}
\usage{
green
}
\description{
A hex code for NACCHO green
}
\keyword{datasets}
