% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{hist_theme}
\alias{hist_theme}
\title{NACCHO Histogram Theme}
\format{
\subsection{\code{hist_theme}}{

A list with a length of 12
}
}
\source{
adapted from the NACCHO basic theme, remove bar spacing
}
\usage{
hist_theme
}
\description{
A list containing NACCHO's histogram theme
}
\keyword{datasets}
