% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{histogram}
\alias{histogram}
\title{NACCHO Histogram Theme}
\format{
\subsection{\code{histogram}}{

A list with a length of 12
}
}
\source{
adapted from the NACCHO basic theme, remove bar spacing
}
\usage{
histogram
}
\description{
A list containing NACCHO's histogram theme
}
\keyword{datasets}
