% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{light}
\alias{light}
\title{NACCHO Off White}
\format{
\subsection{\code{light}}{

A single vector as a hex code
}
}
\source{
adapted from the NACCHO style guide for greater accessibility
}
\usage{
light
}
\description{
A hex code for NACCHO offwhite
}
\keyword{datasets}
