% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{line}
\alias{line}
\title{NACCHO Line Theme}
\format{
\subsection{\code{line}}{

A list with a length of 12
}
}
\source{
adapted from the NACCHO style guide for line graphs
}
\usage{
line
}
\description{
A list containing NACCHO's line highcharter theme for line graphs
}
\keyword{datasets}
