% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{minimal}
\alias{minimal}
\title{NACCHO Minimal Theme}
\format{
\subsection{\code{minimal}}{

A list with a length of 11
}
}
\source{
adapted from the NACCHO style guide for greater accessibility
}
\usage{
minimal
}
\description{
A list containing NACCHO's minimal highcharter theme
}
\keyword{datasets}
