% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{naccho_bare_theme}
\alias{naccho_bare_theme}
\title{NACCHO Bare  Theme}
\format{
\subsection{\code{naccho_bare_theme}}{

A list with a length of 12
}
}
\source{
adapted from the NACCHO style guide for bare graphs
}
\usage{
naccho_bare_theme
}
\description{
A list containing NACCHO's bare highcharter theme for bare graphs
}
\keyword{datasets}
