% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{naccho_blue}
\alias{naccho_blue}
\title{NACCHO Blue}
\format{
\subsection{\code{naccho_blue}}{

A single vector as a hex code
}
}
\source{
adapted from the NACCHO style guide for greater accessibility
}
\usage{
naccho_blue
}
\description{
A hex code for NACCHO blue
}
\keyword{datasets}
