% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{naccho_colors}
\alias{naccho_colors}
\title{NACCHO Main Colors}
\format{
\subsection{\code{naccho_colors}}{

A vector containing 5 objects
}
}
\source{
adapted from the NACCHO style guide for greater accessibility
}
\usage{
naccho_colors
}
\description{
A vector of 5 objects with NACCHO's main colors for visualizations
}
\keyword{datasets}
