% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{naccho_grey}
\alias{naccho_grey}
\title{NACCHO Grey}
\format{
\subsection{\code{naccho_grey}}{

A single vector as a hex code
}
}
\source{
adapted from the NACCHO style guide for greater accessibility
}
\usage{
naccho_grey
}
\description{
A hex code for NACCHO grey
}
\keyword{datasets}
