% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{naccho_line_theme}
\alias{naccho_line_theme}
\title{NACCHO Line Theme}
\format{
\subsection{\code{naccho_bare_theme}}{

A list with a length of 12
}
}
\source{
adapted from the NACCHO style guide for line graphs
}
\usage{
naccho_line_theme
}
\description{
A list containing NACCHO's line highcharter theme for line graphs
}
\keyword{datasets}
