% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{naccho_minimal_theme}
\alias{naccho_minimal_theme}
\title{NACCHO Minimal Theme}
\format{
\subsection{\code{naccho_minimal_theme}}{

A list with a length of 11
}
}
\source{
adapted from the NACCHO style guide for greater accessibility
}
\usage{
naccho_minimal_theme
}
\description{
A list containing NACCHO's minimal highcharter theme
}
\keyword{datasets}
