% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{naccho_minimalTick_theme}
\alias{naccho_minimalTick_theme}
\title{NACCHO Minimal Tick Theme}
\format{
\subsection{\code{naccho_minimalTick_theme}}{

A list with a length of 12
}
}
\source{
adapted from the NACCHO style guide and made only for use with tick bars
}
\usage{
naccho_minimalTick_theme
}
\description{
A list containing NACCHO's minimal highcharter theme for tick bars
}
\keyword{datasets}
