% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{naccho_orange}
\alias{naccho_orange}
\title{NACCHO Orange}
\format{
\subsection{\code{naccho_orange}}{

A single vector as a hex code
}
}
\source{
adapted from the NACCHO style guide for greater accessibility
}
\usage{
naccho_orange
}
\description{
A hex code for NACCHO orange
}
\keyword{datasets}
