% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{pink}
\alias{pink}
\title{NACCHO Pink}
\format{
\subsection{\code{pink}}{

A single vector as a hex code
}
}
\source{
adapted from the NACCHO style guide for greater accessibility
}
\usage{
pink
}
\description{
A hex code for NACCHO pink
}
\keyword{datasets}
