% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{naccho_teal}
\alias{naccho_teal}
\title{NACCHO Teal}
\format{
\subsection{\code{naccho_teal}}{

A single vector as a hex code
}
}
\source{
adapted from the NACCHO style guide for greater accessibility
}
\usage{
naccho_teal
}
\description{
A hex code for NACCHO teal
}
\keyword{datasets}
