% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{navy}
\alias{navy}
\title{NACCHO Blue}
\format{
\subsection{\code{navy}}{

A single vector as a hex code
}
}
\source{
adapted from the NACCHO style guide for greater accessibility
}
\usage{
navy
}
\description{
A hex code for NACCHO blue
}
\keyword{datasets}
