% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{scatter}
\alias{scatter}
\title{NACCHO Scatter Plot Theme}
\format{
\subsection{\code{butterfly}}{

A list with a length of 13
}
}
\source{
NACCHO
}
\usage{
scatter
}
\description{
A list containing NACCHO's bar graph theme with a plotline at "0"
}
\keyword{datasets}
