% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{terracotta}
\alias{terracotta}
\title{NACCHO Orange}
\format{
\subsection{\code{terracotta}}{

A single vector as a hex code
}
}
\source{
adapted from the NACCHO style guide for greater accessibility
}
\usage{
terracotta
}
\description{
A hex code for NACCHO orange
}
\keyword{datasets}
