% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{tick}
\alias{tick}
\title{NACCHO Minimal Tick Theme}
\format{
\subsection{\code{tick}}{

A list with a length of 12
}
}
\source{
adapted from the NACCHO style guide and made only for use with tick bars
}
\usage{
tick
}
\description{
A list containing NACCHO's minimal highcharter theme for tick bars
}
\keyword{datasets}
