% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{naccho_pink}
\alias{naccho_pink}
\title{NACCHO Pink}
\format{
\subsection{\code{naccho_pink}}{

A single vector as a hex code
}
}
\source{
adapted from the NACCHO style guide for greater accessibility
}
\usage{
naccho_pink
}
\description{
A hex code for NACCHO pink
}
\keyword{datasets}
